Refactor+Read.me config+BusLine/StopId Dropdown
