+++ src/app/layouts/BasicFormLayout.tsx
@@ -1,5 +1,5 @@
 import { ReactNode } from "react";
-import Button from "../components/Button";
+import Button from "../../components/Button";
 
 interface IBasicFormLayout {
   title: string;
